+++ examples/boxes/main.cpp
@@ -39,6 +39,7 @@
 #include <osgShadow/ShadowMap>
 
 using namespace dart;
+using dart::simulation::CollisionDetectorType;
 
 [[nodiscard]] dynamics::SkeletonPtr createBox(
     const Eigen::Vector3d& position,
@@ -75,11 +76,9 @@ using namespace dart;
 int main()
 {
   // Create an empty world
-  auto world = simulation::World::create();
-
-  // Set collision detector type
-  world->getConstraintSolver()->setCollisionDetector(
-      collision::BulletCollisionDetector::create());
+  auto world = simulation::World::create(simulation::WorldConfig{
+      .collisionDetector = CollisionDetectorType::Bullet,
+  });
 
   // Create dim x dim x dim boxes
   auto dim = 5;