+++ tests/integration/collision/test_CollisionAccuracy.cpp
@@ -48,6 +48,7 @@
 #include <gtest/gtest.h>
 
 using namespace dart::test;
+using dart::simulation::CollisionDetectorType;
 
 //==============================================================================
 TEST(Issue1184, Accuracy)
@@ -104,9 +105,11 @@ TEST(Issue1184, Accuracy)
     for (const auto& objectShapeFunction : objectShapeFunctions) {
       for (const double halfsize : halfsizes) {
         for (const bool falling : fallingModes) {
-          auto world = dart::simulation::World::create("test");
-          world->getConstraintSolver()->setCollisionDetector(
-              dart::collision::BulletCollisionDetector::create());
+          auto world
+              = dart::simulation::World::create(dart::simulation::WorldConfig{
+                  .name = "test",
+                  .collisionDetector = CollisionDetectorType::Bullet,
+              });
 
           Eigen::Isometry3d tf_object = Eigen::Isometry3d::Identity();
           const double initialHeight