+++ tests/integration/constraint/test_Constraint.cpp
@@ -124,12 +124,12 @@ void ConstraintTest::SingleContactTest(const std::string& /*_fileName*/)
   // std::size_t testCount = 1;
 #endif
 
-  WorldPtr world = World::create();
+  WorldPtr world = World::create(WorldConfig{
+      .collisionDetector = CollisionDetectorType::Dart,
+  });
   EXPECT_TRUE(world != nullptr);
   world->setGravity(Vector3d(0.0, -10.00, 0.0));
   world->setTimeStep(0.001);
-  world->getConstraintSolver()->setCollisionDetector(
-      DARTCollisionDetector::create());
 
   SkeletonPtr sphereSkel = createSphere(0.05, Vector3d(0.0, 1.0, 0.0));
   BodyNode* sphere = sphereSkel->getBodyNode(0);